--- downloaded_data/farm_gate_agriculture_energy/2023-12-13/farm_gate_agriculture_energy.zip
@@ -1,1 +1,0 @@
-../../../.git/annex/objects/4m/p0/MD5E-s1131872--3a3329f2115c62bab08ba71183623db7.zip/MD5E-s1131872--3a3329f2115c62bab08ba71183623db7.zip
